# Change page projection to always process columnar