--- a/presto-main/src/main/java/com/facebook/presto/SystemSessionProperties.java
+++ b/presto-main/src/main/java/com/facebook/presto/SystemSessionProperties.java
@@ -54,7 +54,6 @@ public final class SystemSessionProperties
     public static final String REDISTRIBUTE_WRITES = "redistribute_writes";
     public static final String PUSH_TABLE_WRITE_THROUGH_UNION = "push_table_write_through_union";
     public static final String EXECUTION_POLICY = "execution_policy";
-    public static final String PROCESSING_OPTIMIZATION = "processing_optimization";
     public static final String DICTIONARY_AGGREGATION = "dictionary_aggregation";
     public static final String PLAN_WITH_TABLE_NODE_PARTITIONING = "plan_with_table_node_partitioning";
     public static final String COLOCATED_JOIN = "colocated_join";
@@ -197,11 +196,6 @@ public final class SystemSessionProperties
                         RESOURCE_OVERCOMMIT,
                         "Use resources which are not guaranteed to be available to the query",
                         false,
-                        false),
-                stringSessionProperty(
-                        PROCESSING_OPTIMIZATION,
-                        "Type of optimization for query processing",
-                        featuresConfig.getProcessingOptimization(),
                         false),
                 booleanSessionProperty(
                         DICTIONARY_AGGREGATION,
@@ -367,11 +361,6 @@ public final class SystemSessionProperties
     public static boolean isShareIndexLoading(Session session)
     {
         return session.getSystemProperty(TASK_SHARE_INDEX_LOADING, Boolean.class);
-    }
-
-    public static String getProcessingOptimization(Session session)
-    {
-        return session.getSystemProperty(PROCESSING_OPTIMIZATION, String.class);
     }
 
     public static boolean isDictionaryAggregationEnabled(Session session)