--- a/presto-main/src/main/java/com/facebook/presto/operator/FilterAndProjectOperator.java
+++ b/presto-main/src/main/java/com/facebook/presto/operator/FilterAndProjectOperator.java
@@ -16,14 +16,12 @@ package com.facebook.presto.operator;
 import com.facebook.presto.spi.Page;
 import com.facebook.presto.spi.PageBuilder;
 import com.facebook.presto.spi.type.Type;
-import com.facebook.presto.sql.analyzer.FeaturesConfig;
 import com.facebook.presto.sql.planner.plan.PlanNodeId;
 import com.google.common.collect.ImmutableList;
 
 import java.util.List;
 import java.util.function.Supplier;
 
-import static com.facebook.presto.SystemSessionProperties.getProcessingOptimization;
 import static com.google.common.base.Preconditions.checkState;
 import static java.util.Objects.requireNonNull;
 
@@ -35,9 +33,7 @@ public class FilterAndProjectOperator
 
     private final PageBuilder pageBuilder;
     private final PageProcessor processor;
-    private final String processingOptimization;
     private Page currentPage;
-    private int currentPosition;
     private boolean finishing;
 
     public FilterAndProjectOperator(OperatorContext operatorContext, Iterable<? extends Type> types, PageProcessor processor)
@@ -45,7 +41,6 @@ public class FilterAndProjectOperator
         this.processor = requireNonNull(processor, "processor is null");
         this.operatorContext = requireNonNull(operatorContext, "operatorContext is null");
         this.types = ImmutableList.copyOf(requireNonNull(types, "types is null"));
-        this.processingOptimization = getProcessingOptimization(operatorContext.getSession());
         this.pageBuilder = new PageBuilder(getTypes());
     }
 
@@ -87,37 +82,15 @@ public class FilterAndProjectOperator
         checkState(!pageBuilder.isFull(), "Page buffer is full");
 
         currentPage = page;
-        currentPosition = 0;
     }
 
     @Override
     public final Page getOutput()
     {
         if (!pageBuilder.isFull() && currentPage != null) {
-            switch (processingOptimization) {
-                case FeaturesConfig.ProcessingOptimization.COLUMNAR: {
-                    Page page = processor.processColumnar(operatorContext.getSession().toConnectorSession(), currentPage, getTypes());
-                    currentPage = null;
-                    currentPosition = 0;
-                    return page;
-                }
-                case FeaturesConfig.ProcessingOptimization.COLUMNAR_DICTIONARY: {
-                    Page page = processor.processColumnarDictionary(operatorContext.getSession().toConnectorSession(), currentPage, getTypes());
-                    currentPage = null;
-                    currentPosition = 0;
-                    return page;
-                }
-                case FeaturesConfig.ProcessingOptimization.DISABLED: {
-                    currentPosition = processor.process(operatorContext.getSession().toConnectorSession(), currentPage, currentPosition, currentPage.getPositionCount(), pageBuilder);
-                    if (currentPosition == currentPage.getPositionCount()) {
-                        currentPage = null;
-                        currentPosition = 0;
-                    }
-                    break;
-                }
-                default:
-                    throw new IllegalStateException();
-            }
+            Page page = processor.process(operatorContext.getSession().toConnectorSession(), currentPage, getTypes());
+            currentPage = null;
+            return page;
         }
 
         if (!finishing && !pageBuilder.isFull() || pageBuilder.isEmpty()) {