--- a/presto-main/src/main/java/com/facebook/presto/operator/FilterAndProjectOperator.java
+++ b/presto-main/src/main/java/com/facebook/presto/operator/FilterAndProjectOperator.java
@@ -13,8 +13,10 @@
  */
 package com.facebook.presto.operator;
 
+import com.facebook.presto.memory.LocalMemoryContext;
+import com.facebook.presto.operator.project.PageProcessor;
+import com.facebook.presto.operator.project.PageProcessorOutput;
 import com.facebook.presto.spi.Page;
-import com.facebook.presto.spi.PageBuilder;
 import com.facebook.presto.spi.type.Type;
 import com.facebook.presto.sql.planner.plan.PlanNodeId;
 import com.google.common.collect.ImmutableList;
@@ -22,6 +24,7 @@ import com.google.common.collect.ImmutableList;
 import java.util.List;
 import java.util.function.Supplier;
 
+import static com.facebook.presto.operator.project.PageProcessorOutput.EMPTY_PAGE_PROCESSOR_OUTPUT;
 import static com.google.common.base.Preconditions.checkState;
 import static java.util.Objects.requireNonNull;
 
@@ -30,18 +33,18 @@ public class FilterAndProjectOperator
 {
     private final OperatorContext operatorContext;
     private final List<Type> types;
+    private final LocalMemoryContext outputMemoryContext;
 
-    private final PageBuilder pageBuilder;
     private final PageProcessor processor;
-    private Page currentPage;
+    private PageProcessorOutput currentOutput = EMPTY_PAGE_PROCESSOR_OUTPUT;
     private boolean finishing;
 
     public FilterAndProjectOperator(OperatorContext operatorContext, Iterable<? extends Type> types, PageProcessor processor)
     {
         this.processor = requireNonNull(processor, "processor is null");
         this.operatorContext = requireNonNull(operatorContext, "operatorContext is null");
+        this.outputMemoryContext = operatorContext.getSystemMemoryContext().newLocalMemoryContext();
         this.types = ImmutableList.copyOf(requireNonNull(types, "types is null"));
-        this.pageBuilder = new PageBuilder(getTypes());
     }
 
     @Override
@@ -65,13 +68,18 @@ public class FilterAndProjectOperator
     @Override
     public final boolean isFinished()
     {
-        return finishing && pageBuilder.isEmpty() && currentPage == null;
+        boolean finished = finishing && !currentOutput.hasNext();
+        if (finished) {
+            currentOutput = EMPTY_PAGE_PROCESSOR_OUTPUT;
+            outputMemoryContext.setBytes(0);
+        }
+        return finished;
     }
 
     @Override
     public final boolean needsInput()
     {
-        return !finishing && !pageBuilder.isFull() && currentPage == null;
+        return !finishing && !currentOutput.hasNext();
     }
 
     @Override
@@ -79,27 +87,19 @@ public class FilterAndProjectOperator
     {
         checkState(!finishing, "Operator is already finishing");
         requireNonNull(page, "page is null");
-        checkState(!pageBuilder.isFull(), "Page buffer is full");
+        checkState(!currentOutput.hasNext(), "Page buffer is full");
 
-        currentPage = page;
+        currentOutput = processor.process(operatorContext.getSession().toConnectorSession(), page);
+        outputMemoryContext.setBytes(currentOutput.getRetainedSizeInBytes());
     }
 
     @Override
     public final Page getOutput()
     {
-        if (!pageBuilder.isFull() && currentPage != null) {
-            Page page = processor.process(operatorContext.getSession().toConnectorSession(), currentPage, getTypes());
-            currentPage = null;
-            return page;
-        }
-
-        if (!finishing && !pageBuilder.isFull() || pageBuilder.isEmpty()) {
+        if (!currentOutput.hasNext()) {
             return null;
         }
-
-        Page page = pageBuilder.build();
-        pageBuilder.reset();
-        return page;
+        return currentOutput.next();
     }
 
     public static class FilterAndProjectOperatorFactory
@@ -115,8 +115,8 @@ public class FilterAndProjectOperator
         {
             this.operatorId = operatorId;
             this.planNodeId = requireNonNull(planNodeId, "planNodeId is null");
-            this.processor = processor;
-            this.types = types;
+            this.processor = requireNonNull(processor, "processor is null");
+            this.types = ImmutableList.copyOf(requireNonNull(types, "types is null"));
         }
 
         @Override