… presto-benchmark/src/main/java/com/facebook/presto/benchmark/AbstractOperatorBenchmark.java
… presto-benchmark/src/main/java/com/facebook/presto/benchmark/HandTpchQuery6.java
… presto-benchmark/src/main/java/com/facebook/presto/benchmark/PredicateFilterBenchmark.java
… presto-hive/src/test/java/com/facebook/presto/hive/TestOrcPageSourceMemoryTracking.java
… presto-main/src/main/java/com/facebook/presto/operator/FilterFunction.java
… presto-main/src/main/java/com/facebook/presto/operator/FilterFunctions.java
… presto-main/src/main/java/com/facebook/presto/operator/GenericCursorProcessor.java
… presto-main/src/main/java/com/facebook/presto/operator/ProjectionFunctions.java
… presto-main/src/main/java/com/facebook/presto/operator/index/TupleFilterProcessor.java
… presto-main/src/main/java/com/facebook/presto/operator/project/ConstantPageProjection.java
… presto-main/src/main/java/com/facebook/presto/operator/project/DictionaryAwarePageFilter.java
… presto-main/src/main/java/com/facebook/presto/operator/project/DictionaryAwarePageProjection.java
… presto-main/src/main/java/com/facebook/presto/operator/project/InputChannels.java
… presto-main/src/main/java/com/facebook/presto/operator/project/InputPageProjection.java
… presto-main/src/main/java/com/facebook/presto/operator/project/InterpretedCursorProcessor.java
… presto-main/src/main/java/com/facebook/presto/operator/project/InterpretedPageFilter.java
… presto-main/src/main/java/com/facebook/presto/operator/project/InterpretedPageProjection.java
… presto-main/src/main/java/com/facebook/presto/operator/project/PageFieldsToInputParametersRewriter.java
… presto-main/src/main/java/com/facebook/presto/operator/project/PageFilter.java
… presto-main/src/main/java/com/facebook/presto/operator/project/PageProcessor.java
… presto-main/src/main/java/com/facebook/presto/operator/project/PageProcessorOutput.java
… presto-main/src/main/java/com/facebook/presto/operator/project/PageProjection.java
… presto-main/src/main/java/com/facebook/presto/operator/project/SelectedPositions.java
… presto-main/src/main/java/com/facebook/presto/sql/analyzer/FeaturesConfig.java
… presto-main/src/main/java/com/facebook/presto/sql/gen/BytecodeUtils.java
… presto-main/src/main/java/com/facebook/presto/sql/gen/ExpressionCompiler.java
… presto-main/src/main/java/com/facebook/presto/sql/gen/PageFunctionCompiler.java
… presto-main/src/main/java/com/facebook/presto/sql/gen/PageProcessorCompiler.java
… presto-main/src/main/java/com/facebook/presto/sql/gen/VarArgsToMapAdapterGenerator.java
… presto-main/src/main/java/com/facebook/presto/sql/planner/ExpressionInterpreter.java