- Dynamically size projection batch size based on previous output
- Dynamically enable/disable dictionary processing based on previous effectiveness
- Add fast path for constant projection
- Add fast path for identity projection
- Skip filtering for TRUE filter
- Change block builders to lazily allocate initial buffers to improve memory efficiency
- Remove processing_optimization session property and optimizer.processing-optimization configuration property

# Performance
This change should not improve or hurt performance of columnar or dictionary, but simply makes it possible to enable them at all times (no configuration).  I ran the following benchmarks to verify this:

## Before
```
Benchmark                          Mode  Cnt     Score    Error  Units
BenchmarkPageProcessor.compiled   thrpt   50  4399.932 ± 66.100  ops/s
BenchmarkPageProcessor.handCoded  thrpt   50  4686.470 ± 76.373  ops/s

Benchmark                                         (name)  Mode  Cnt    Score   Error  Units
BenchmarkArrayDistinct.arrayDistinct      array_distinct  avgt   20   44.867 ± 0.957  ns/op
BenchmarkArrayDistinct.arrayDistinct  old_array_distinct  avgt   20  163.298 ± 4.158  ns/op

Benchmark                             (name)  Mode  Cnt    Score   Error  Units
BenchmarkArrayFilter.benchmark        filter  avgt   20  120.572 ± 5.021  ns/op
BenchmarkArrayFilter.benchmark  exact_filter  avgt   20   76.784 ± 3.005  ns/op

Benchmark                     Mode  Cnt    Score   Error  Units
BenchmarkArrayJoin.benchmark  avgt   60  186.730 ± 3.446  ns/op

Benchmark                             (name)  Mode  Cnt    Score   Error  Units
BenchmarkArraySort.arraySort      array_sort  avgt   20   99.980 ± 8.269  ns/op

Benchmark                     (mapConfig)  Mode  Cnt     Score    Error  Units
BenchmarkMapConcat.mapConcat   left_empty  avgt   20   265.374 ±  4.648  ns/op
BenchmarkMapConcat.mapConcat  right_empty  avgt   20   257.638 ±  3.667  ns/op
BenchmarkMapConcat.mapConcat   both_empty  avgt   20    38.838 ±  1.135  ns/op
BenchmarkMapConcat.mapConcat    non_empty  avgt   20  1817.320 ± 70.188  ns/op

Benchmark                           (mapSize)      (name)  Mode  Cnt     Score    Error  Units
BenchmarkMapSubscript.mapSubscript          1   fix-width  avgt   20    45.505 ±  7.904  ns/op
BenchmarkMapSubscript.mapSubscript          1   var-width  avgt   20    84.766 ±  1.569  ns/op
BenchmarkMapSubscript.mapSubscript          1  dictionary  avgt   20    67.824 ±  5.565  ns/op
BenchmarkMapSubscript.mapSubscript         13   fix-width  avgt   20  1481.749 ± 18.282  ns/op
BenchmarkMapSubscript.mapSubscript         13   var-width  avgt   20  2538.951 ± 42.323  ns/op
BenchmarkMapSubscript.mapSubscript         13  dictionary  avgt   20  2167.449 ± 42.817  ns/op

```

## After
```
Benchmark                          Mode  Cnt     Score     Error  Units
BenchmarkPageProcessor.compiled   thrpt   50  4367.725 ± 171.266  ops/s
BenchmarkPageProcessor.handCoded  thrpt   50  4522.360 ±  94.679  ops/s

Benchmark                                         (name)  Mode  Cnt    Score   Error  Units
BenchmarkArrayDistinct.arrayDistinct      array_distinct  avgt   20   45.514 ± 2.885  ns/op
BenchmarkArrayDistinct.arrayDistinct  old_array_distinct  avgt   20  160.116 ± 3.091  ns/op

Benchmark                             (name)  Mode  Cnt    Score   Error  Units
BenchmarkArrayFilter.benchmark        filter  avgt   20  123.771 ± 4.506  ns/op
BenchmarkArrayFilter.benchmark  exact_filter  avgt   20   77.407 ± 1.768  ns/op

Benchmark                     Mode  Cnt    Score   Error  Units
BenchmarkArrayJoin.benchmark  avgt   60  183.565 ± 3.695  ns/op

Benchmark                             (name)  Mode  Cnt    Score   Error  Units
BenchmarkArraySort.arraySort      array_sort  avgt   20   87.615 ± 4.817  ns/op

Benchmark                     (mapConfig)  Mode  Cnt     Score    Error  Units
BenchmarkMapConcat.mapConcat   left_empty  avgt   20   259.042 ±  6.192  ns/op
BenchmarkMapConcat.mapConcat  right_empty  avgt   20   261.633 ±  7.658  ns/op
BenchmarkMapConcat.mapConcat   both_empty  avgt   20    38.872 ±  0.308  ns/op
BenchmarkMapConcat.mapConcat    non_empty  avgt   20  1690.355 ± 34.162  ns/op

Benchmark                           (mapSize)      (name)  Mode  Cnt     Score    Error  Units
BenchmarkMapSubscript.mapSubscript          1   fix-width  avgt   20    37.986 ±  0.705  ns/op
BenchmarkMapSubscript.mapSubscript          1   var-width  avgt   20    82.147 ±  2.596  ns/op
BenchmarkMapSubscript.mapSubscript          1  dictionary  avgt   20    60.646 ±  1.504  ns/op
BenchmarkMapSubscript.mapSubscript         13   fix-width  avgt   20  1517.320 ± 37.219  ns/op
BenchmarkMapSubscript.mapSubscript         13   var-width  avgt   20  2365.492 ± 88.507  ns/op
BenchmarkMapSubscript.mapSubscript         13  dictionary  avgt   20  2087.881 ± 19.953  ns/op
```