@losipiuk 2017-03-11
This is huge so I will review commit by commit.

Btw. Commit order on github is messed up as it sorts the commits by "CommitDate" instead "AuthorDate".
To workaround it before pushing a branch to be a PR we do:
`git rebase -i --exec "sleep 1" origin/master" and slightly change the commit message of first commit.

Review progress (to be updated as review goes on): 

* (TBD) 268c542 - Improve BlockAssertions.assertBlockEquals check
* (TBD) b388fc4 - Change block builders to lazily allocate initial buffers
* (TBD) bf1c463 - Fix IntelliJ warnings
* (TBD) d9aa6ff - Minor cleanup of projection classes
* (TBD) 416b19c - Update benchmarks to use columnar processing
* (TBD) d79fdf1 - Use RowExpression factory methods instead of constructor
* (TBD) 3c95c13 - Remove processing_optimization
* (TBD) 10d9edd - Make filter optional in ExpressionCompiler
* (TBD) d0328b1 - Change PageProcessor to allow partial results
* (TBD) f908567 - Rewrite columnar page processor
* (TBD) 9275e65 - Add dictionary aware filter and project