@losipiuk 2017-03-14
@dain I went through everything. Some comments but nothing major. 
Code is for sure non-trivial so some excuse me if some comments does not make sense :). 

Note that for some reason GitHub renders some of comments as "outdated" and those are not visible if you click on individual commits from PR, so you have to access them via "Conversations" tab.

As you go through comments it would be helpful if you marked the already addressed with emoticon or just a reply.