@losipiuk 2017-04-13
Seems fine to me. Let's not keep it hanging around to avoid painful rebasing.